requirements.txt
python-quickbooks
python-dotenv
requests
flask
google-generativeai
